+++ Acidmanic.Utilities.Reflection/ObjectInstantiator.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Reflection;
 using Acidmanic.Utilities.Reflection.Attributes;
 using Acidmanic.Utilities.Reflection.Extensions;
+using Acidmanic.Utilities.Reflection.ObjectTree;
 
 namespace Acidmanic.Utilities.Reflection
 {
@@ -29,8 +30,70 @@ namespace Acidmanic.Utilities.Reflection
 
         public object BlindInstantiate(Type type)
         {
-            var constructor = FindConcreteConstructor(type);
+            if (TypeCheck.IsEffectivelyPrimitive(type))
+            {
+                return InstantiatePrimitive(type);
+            }
 
+            Func<object> constructor = null;
+            
+            
+            if (type.IsAbstract || type.IsInterface)
+            {
+                constructor = FindConcreteConstructor(type);
+            }
+            else
+            {
+                constructor = FindPrimitiveConstructor(type);
+            }
+
+            object createdObject = null;
+
+            try
+            {
+                createdObject = constructor.Invoke();
+            }
+            catch (Exception _)
+            {
+            }
+
+            // Check for cases that a wrong concrete class has been found
+            if (createdObject != null && !type.IsInstanceOfType(createdObject))
+            {
+                createdObject = null;
+            }
+
+            return createdObject;
+        }
+
+
+        private Func<object> FindPrimitiveConstructor(Type type)
+        {
+            var constructors = type.GetConstructors()
+                .OrderBy(c => c.GetParameters().Length);
+
+            foreach (var constructor in constructors)
+            {
+                if (IsFullyPrimitive(constructor))
+                {
+                    var parameters = constructor.GetParameters();
+
+                    var values = new object[parameters.Length];
+
+                    for (int i = 0; i < parameters.Length; i++)
+                    {
+                        values[i] = InstantiatePrimitive(parameters[i].ParameterType);
+                    }
+
+                    return () => constructor.Invoke(values);
+                }
+            }
+
+            return null;
+        }
+
+        private object InstantiatePrimitive(Type type)
+        {
             if (type == typeof(string))
             {
                 return default(string);
@@ -41,8 +104,24 @@ namespace Acidmanic.Utilities.Reflection
                 return Activator.CreateInstance(type);
             }
 
-            return constructor.Invoke();
+            return default;
         }
+
+        private bool IsFullyPrimitive(ConstructorInfo constructor)
+        {
+            var parameters = constructor.GetParameters();
+
+            foreach (var parameter in parameters)
+            {
+                if (!TypeCheck.IsEffectivelyPrimitive(parameter.ParameterType))
+                {
+                    return false;
+                }
+            }
+
+            return true;
+        }
+
 
         private Func<object> FindConcreteConstructor(Type type)
         {